Rename SerializeX to SerializedX
